Address compilation warnings

+++ fdbserver/workloads/BackupCorrectnessPartitioned.actor.cpp
@@ -682,10 +682,10 @@ struct BackupAndRestorePartitionedCorrectnessWorkload : TestWorkload {
 					Standalone<StringRef> restoreTag(self->backupTag.toString() + "_" + std::to_string(restoreIndex));
 					restoreTags.push_back(restoreTag);
 					printf("BackupCorrectness, backupAgent.restore is called for restoreIndex:%d tag:%s "
-					       "TargetVersion:%d\n",
+					       "TargetVersion:%lld\n",
 					       restoreIndex,
 					       restoreTag.toString().c_str(),
-					       targetVersion);
+					       static_cast<long long>(targetVersion));
 					restores.push_back(backupAgent.restore(cx,
 					                                       cx,
 					                                       restoreTag,
